requirements.txt
Django==3.2.8
djangorestframework==3.12.4
psycopg2-binary==2.9.1
PyPDF2==1.26.0
pdfminer.six==20201018
spacy==3.1.3
tensorflow==2.6.0
transformers==4.10.3
celery==5.1.2
pytest==6.2.5
pytest-django==4.4.0
gunicorn==20.1.0
django-cors-headers==3.10.0
django-filter==2.4.0
django-extensions==3.1.3
django-environ==0.4.5
django-storages==1.11.1
boto3==1.18.33
redis==3.5.3
flower==1.0.0
django-celery-results==2.0.1
django-celery-beat==2.2.1
django-health-check==3.16.4
django-prometheus==2.1.0
django-redis==5.0.0
django-rest-knox==4.1.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
python-decouple==3.4
python-dotenv==0.19.0
whitenoise==5.3.0
Pillow==8.3.2
python-magic==0.4.24
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==3.4.8
Faker==8.10.0
factory-boy==3.2.0
coverage==5.5
codecov==2.1.12
flake8==3.9.2
isort==5.9.3
black==21.9b0
mypy==0.910
pre-commit==2.15.0
bandit==1.7.0
safety==1.10.3
twine==3.4.2
wheel==0.37.0
setuptools==58.0.4
docker==5.0.0
docker-compose==1.29.2
kubernetes==18.20.0
